Moved read_pattern into decode_core
