+++ src/decode_core.rs
@@ -60,6 +60,27 @@ pub struct DecodeFailsBreakdown {
     pub parity_blocks_decode_failed: u64,
 }
 
+pub enum ReadPattern {
+    Sequential(Option<(usize, usize, usize)>),
+    BurstErrorResistant(usize, usize, usize),
+}
+
+impl ReadPattern {
+    pub fn new(
+        from_pos: Option<u64>,
+        to_pos: Option<RangeEnd<u64>>,
+        data_par_burst: Option<(usize, usize, usize)>,
+    ) -> Self {
+        match data_par_burst {
+            Some((data, parity, burst)) => match (from_pos, to_pos) {
+                (None, None) => ReadPattern::BurstErrorResistant(data, parity, burst),
+                _ => ReadPattern::Sequential(Some((data, parity, burst))),
+            },
+            None => ReadPattern::Sequential(None),
+        }
+    }
+}
+
 #[derive(Clone, Debug)]
 pub enum DecodeFailStats {
     Breakdown(DecodeFailsBreakdown),
